move all tests to TensorFlowRuntime
