+++ test/TensorFlowRuntime/numpy_conversion.swift
@@ -58,6 +58,12 @@ NumpyConversionTests.test("shaped-array-conversion") {
                 array)
   }
 
+  let numpyArray1D = np.ones(28)
+  let reshaped3D = np.reshape(numpyArray1D, [2, 7, 2] as TensorShape)
+  expectEqual(TensorShape(reshaped3D.shape), [2, 7, 2])
+  let reshaped2D = np.reshape(reshaped3D, [14, 2] as TensorShape)
+  expectEqual(TensorShape(reshaped2D.shape), [14, 2])
+
   let numpyArrayStrided = np.array([[1, 2], [1, 2]], dtype: np.int32)[
       Python.slice(Python.None), 1]
   // Assert that the array has a stride, so that we're certainly testing a
